make sure Decoder raises DecodeError or DataViewIndexOutOfBoundsError
